requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
pip==22.2.2
setuptools==65.3.0
sqlparse==0.4.2
wheel==0.37.1
djangorestframework
rest_condition
django-cors-headers
django-filter
django-rest-auth
djangorestframework-simplejwt